% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beds}
\alias{beds}
\title{#' 10 observations showing patient time in and out.}
\format{
A data frame with 10 rows and 4 variables:
\describe{
  \item{bed}{a grouping variable, representing a specific bed}
  \item{patient}{individual patients}
  \item{start_time}{time each patient moved into their current bed}
  \item{end_time}{time each patient moved out of their current bed}
}
}
\usage{
beds
}
\description{
A dataset containing ten sets of time in and out by bed.
This simulates patients moving in and out of departments or hospitals.
}
\keyword{datasets}
