% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_census.R
\name{interval_census}
\alias{interval_census}
\title{interval_census
A census of the number of patients , or moves/ transfers, by interval}
\usage{
interval_census(
  df,
  identifier,
  admit,
  discharge,
  group_var,
  time_unit = "1 hour",
  results = c("patient", "group", "total"),
  uniques = TRUE,
  timezone = NULL
)
}
\arguments{
\item{df}{dataframe, tibble or data.table}

\item{identifier}{unique patient identifier}

\item{admit}{datetime of admission as POSIXct yyyy-mm-dd hh:mm:ss}

\item{discharge}{datetime of discharge as POSIXct yyyy-mm-dd hh:mm:ss}

\item{group_var}{unique character vector to identify location/clinician at each move}

\item{time_unit}{character string to denote time intervals to count by e.g. "1 hour", "15 mins"}

\item{results}{patient returns one row per patient, groupvar and interval.
group provides an overall grouped count of patients by the specified time interval.
total returns the grand total of patients 'IN' by each unique time interval.}

\item{uniques}{TRUE will only count patients once if they have more than one move
during an interval. Use this to get a distinct count of patients per interval

FALSE will count each patient move, so if a patient moves once during an interval there
will be two rows for tha patient. This is useful if you want to count occupied beds,
or a count of moves / transfers between departments}

\item{timezone}{Your system timezone}
}
\value{
data.table showing identifier, group_var and count by relevant unit of time
}
\description{
interval_census
A census of the number of patients , or moves/ transfers, by interval
}
