% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_time_counter.R
\name{in_time_counter}
\alias{in_time_counter}
\title{in_time_counter}
\usage{
in_time_counter(
  df,
  identifier,
  admit_datetime,
  discharge_datetime,
  group_var,
  time_unit = "1 hour",
  results = c("patient", "group", "total")
)
}
\arguments{
\item{df}{dataframe, tibble or data.table}

\item{identifier}{unique patient identifier}

\item{admit_datetime}{datetime of admission as POSIXct yyyy-mm-dd hh:mm:ss}

\item{discharge_datetime}{datetime of discharge as POSIXct yyyy-mm-dd hh:mm:ss}

\item{group_var}{unique character vector to identify location/clinician at each move}

\item{time_unit}{character string to denote time intervals to count by e.g. "1 hour", "15 mins"}

\item{results}{patient returns one row per patient, groupvar and interval.
group provides an overall grouped count of patients by the specified time interval.
total returns the grand total of patients 'IN' by each unique time interval.}
}
\value{
data.table showing identifier, group_var and count by relevant unit of time
}
\description{
in_time_counter
}
