% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_census.R
\name{interval_census}
\alias{interval_census}
\title{interval_census
A census of the number of patients , or moves/ transfers, by interval}
\usage{
interval_census(
  df,
  identifier,
  admit,
  discharge,
  group_var = NULL,
  time_unit = "1 hour",
  time_adjust_period = NULL,
  time_adjust_value = NULL,
  results = c("patient", "group", "total"),
  uniques = TRUE
)
}
\arguments{
\item{df}{dataframe, tibble or data.table}

\item{identifier}{unique patient identifier}

\item{admit}{datetime of admission as POSIXct yyyy-mm-dd hh:mm:ss}

\item{discharge}{datetime of discharge as POSIXct yyyy-mm-dd hh:mm:ss}

\item{group_var}{unique character vector to identify location/clinician at each move}

\item{time_unit}{character string to denote time intervals to count by e.g. "1 hour", "15 mins"}

\item{time_adjust_period}{"start_sec","start_min","end_sec","end_min"}

\item{time_adjust_value}{integer to adjust the start / end of each period in minutes or seconds}

\item{results}{'patient' returns one row per patient, groupvar and interval.

'group' provides an overall grouped count of patients by the specified time interval.

'total' returns the grand total of patients 'IN' by each unique time interval.}

\item{uniques}{TRUE will count patients once per interval, even if they have
more than one entry per interval, for example, due to move to another location.
Use this to get a *distinct* count of patients per interval

FALSE will count each patient entry per interval.
If a patient moves during an interval there will be  at least two rows for
that patient for that interval.
This is useful if you want to count occupied beds,or a count of moves / transfers between departments}
}
\value{
data.table showing identifier, group variable , and count by relevant unit of time
Also includes the start / end of interval, plus the base date and base hour for
convenient interactive filtering of the results
}
\description{
interval_census
A census of the number of patients , or moves/ transfers, by interval
}
\examples{
\donttest{
interval_census(admit_extract, indentifier ="pat_id", admit = "admit_date",
discharge = "discharge_date", group_var ="location",
time_unit = "1 hour", results = "patient",
uniques = TRUE)
}


}
